# api/openapi.yaml
openapi: 3.1.0
info:
  title: dra-workflow API
  version: 0.1.0
  summary: Workflow-driven, AI-assisted IT service disaster recovery planning and execution.
  description: |
    REST API of the dra-workflow backend (Rust). It implements the data model in
    `docs/domain/dr-plan-model.md` and the workflows in `docs/workflows/`.

    **Hierarchy:** Tenant → IT service → microservice → DR items.

    **Methodology:** NIST SP 800-34 Rev. 1 and BSI-Standard 200-4 (see `docs/domain/standards-mapping.md`).

    ## Conventions
    - **Tenant from the token:** the tenant is **not** part of the path. The backend reads it from the
      `tenant_id` claim of the JWT access token and scopes every query to it. A token without a valid
      tenant claim, or for a tenant the user is not a member of, gets `403`. Ids belonging to another
      tenant return `404`. Only `GET/POST /tenants` (membership list, tenant creation) are tenant-independent.
    - **Collections** are nested under their parent (for example `/services/{serviceId}/microservices`).
      **Single items** are addressed flat by id (for example `/microservices/{microserviceId}`).
    - IDs are UUIDs. Durations are **integer minutes**. Timestamps are RFC 3339 UTC.
    - Updates use `PATCH` with `application/merge-patch+json` (RFC 7396).
    - Optimistic concurrency: every resource has a `version`. Reads return an `ETag`, and `PATCH`/`DELETE`
      accept `If-Match`. A stale version returns `412`.
...
      with `issues`.
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /services:
    get:
      tags: [IT Services]
      operationId: listServices
      summary: List IT services
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: q
          in: query
          schema: { type: string }
        - name: lifecycleStatus
          in: query
          schema: { $ref: '#/components/schemas/ServiceLifecycleStatus' }
        - name: reviewOverdue
          in: query
          description: Only services whose approved plan is past `nextReviewDue`
          schema: { type: boolean }
      responses:
        '200':
          description: IT services
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ItServicePage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [IT Services]
      operationId: createService
      summary: Create an IT service (workflow step 1)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ItServiceCreate' }
      responses:
        '201':
          description: Created
          headers:
            Location: { $ref: '#/components/headers/Location' }
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ItService' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
  /services/{serviceId}:
    parameters:
      - $ref: '#/components/parameters/ServiceId'
    get:
      tags: [IT Services]
      operationId: getService
      summary: Get an IT service
      responses:
        '200':
          description: IT service
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ItService' }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
    patch:
      tags: [IT Services]
      operationId: updateService
      summary: Update an IT service
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: '#/components/schemas/ItServiceFields' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ItService' }
        '404': { $ref: '#/components/responses/NotFound' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
    delete:
      tags: [IT Services]
      operationId: deleteService
      summary: Delete an IT service
      description: Fails with `409` while a recovery run is active. Approved plan versions are kept for audit.
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
...
  /microservices/{microserviceId}/recovery-objectives:
    parameters:
      - $ref: '#/components/parameters/MicroserviceId'
    get:
      tags: [Recovery Objectives]
      operationId: listRecoveryObjectives
      summary: List RTO/RPO objectives of a microservice (workflow step 7)
      responses:
        '200':
          description: Objectives
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/RecoveryObjective' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Recovery Objectives]
      operationId: createRecoveryObjective
      summary: Define an objective (default, or per scenario)
      description: Validated against the service BIA (RTO ≤ service RTO ≤ MTPD, RPO ≤ service RPO).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecoveryObjectiveCreate' }
      responses:
        '201':
          description: Created
          headers:
            Location: { $ref: '#/components/headers/Location' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecoveryObjective' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
  /recovery-objectives/{objectiveId}:
    parameters:
      - $ref: '#/components/parameters/ObjectiveId'
    patch:
      tags: [Recovery Objectives]
      operationId: updateRecoveryObjective
      summary: Update an objective
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: '#/components/schemas/RecoveryObjectiveFields' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecoveryObjective' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
    delete:
      tags: [Recovery Objectives]
      operationId: deleteRecoveryObjective
      summary: Delete an objective
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }

  # ───────────────────────────── Recovery Strategies (Mitigations) ─────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC access token (JWT). Required claims: `sub` (user) and `tenant_id` (UUID of the active tenant).
        The tenant role may come from the token or from the tenant membership. The identity provider is
        still open (docs/overview/open-questions.md).
  headers:
    Location:
      description: URL of the created resource
      schema: { type: string }
    ETag:
      description: Entity tag derived from the resource `version`
      schema: { type: string }
  parameters:
    ServiceId:
      name: serviceId
      in: path
      required: true
      schema: { type: string, format: uuid }
    MicroserviceId:
      name: microserviceId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ObjectiveId:
      name: objectiveId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor from a previous page's `nextCursor`
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    IfMatch:
      name: If-Match
      in: header
      description: ETag of the version being modified (optimistic concurrency)
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Error:
      description: Unexpected error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource not found (or it belongs to a different tenant than the token's)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Conflict:
      description: Conflicts with the current state (for example the resource is still referenced, or the transition is invalid)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    PreconditionFailed:
      description: The `If-Match` version is stale
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationFailed:
      description: Input validation, integrity rule or workflow gate failed
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ValidationProblem' }
...
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details
      properties:
        type: { type: string, format: uri-reference, default: about:blank }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            issues:
              type: array
              items: { $ref: '#/components/schemas/Issue' }
    Issue:
      type: object
      description: A validation error, integrity-rule violation or gate failure
      required: [ruleId, severity, message]
      properties:
        ruleId:
          type: string
          description: Stable rule id, for example `RTO_EXCEEDS_SERVICE_RTO`, `DEPENDENCY_RTO_CONFLICT`, `STEP_WITHOUT_OWNER`, `STRATEGY_GAP`, `ROLE_WITHOUT_DEPUTY`
          examples: [DEPENDENCY_RTO_CONFLICT]
        severity: { type: string, enum: [blocking, warning, info] }
        message: { type: string }
        field: { type: string, description: 'JSON pointer to the offending input field, if any' }
        entityType: { type: string }
        entityId: { type: string, format: uuid }
        workflowStep: { $ref: '#/components/schemas/WorkflowStepKey' }
        standardRef:
          type: string
          description: Related standard reference
          examples: ['NIST SP 800-34 §3.2', 'BSI 200-4 Soll-Ist-Vergleich']
    ResourceMeta:
      type: object
      required: [id, tenantId, version, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        tenantId: { type: string, format: uuid, readOnly: true }
        version: { type: integer, readOnly: true, description: Incremented on every change (optimistic concurrency) }
        createdAt: { type: string, format: date-time, readOnly: true }
        createdBy: { type: string, readOnly: true }
        updatedAt: { type: string, format: date-time, readOnly: true }
        updatedBy: { type: string, readOnly: true }
    ContentMeta:
      type: object
      properties:
        origin:
          type: string
          enum: [user, ai_accepted, import]
          readOnly: true
          description: Provenance of the content
        aiSuggestionId:
          type: [string, 'null']
          format: uuid
          readOnly: true
    PageInfo:
      type: object
      properties:
        nextCursor: { type: [string, 'null'] }
    Minutes:
      type: integer
      minimum: 0
      description: Duration in minutes

    # ─────────── Enums ───────────
...
    ServiceLifecycleStatus:
      type: string
      enum: [planned, active, retired]
    ProtectionRequirement:
      type: string
      description: BSI protection requirement (Schutzbedarf) for availability
      enum: [normal, high, very_high]
    ImpactLevel:
      type: string
      description: NIST FIPS 199 availability impact level
      enum: [low, moderate, high]
...
    WorkflowStepKey:
      type: string
      enum:
        - define_service
        - map_dependencies
        - business_impact
        - brainstorm_scenarios
        - consolidate_scenarios
        - select_scenarios
        - recovery_objectives
        - recovery_strategies
        - runbooks
        - roles
        - communication
        - review_approve
        - test
        - measure
        - improve
...
    PlanStatus:
      type: string
      enum: [draft, in_review, approved, retired]
...
    ItServiceFields:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string }
        businessOwnerId: { type: [string, 'null'], format: uuid, description: Person }
        technicalOwnerId: { type: [string, 'null'], format: uuid, description: Person }
        consumers:
          type: array
          description: Who depends on this service (user groups, business processes)
          items: { type: string }
        protectionRequirementAvailability: { $ref: '#/components/schemas/ProtectionRequirement' }
        impactLevel: { $ref: '#/components/schemas/ImpactLevel' }
        lifecycleStatus: { $ref: '#/components/schemas/ServiceLifecycleStatus' }
    ItServiceCreate:
      allOf:
        - $ref: '#/components/schemas/ItServiceFields'
        - required: [name]
    ItService:
      allOf:
        - $ref: '#/components/schemas/ResourceMeta'
        - $ref: '#/components/schemas/ContentMeta'
        - $ref: '#/components/schemas/ItServiceFields'
        - type: object
          required: [name]
          properties:
            summary: { $ref: '#/components/schemas/ServiceSummary' }
    ServiceSummary:
      type: object
      readOnly: true
      properties:
        microserviceCount: { type: integer }
        selectedScenarioCount: { type: integer }
        workflowCompletionPercent: { type: integer, minimum: 0, maximum: 100 }
        currentPlanVersionId: { type: [string, 'null'], format: uuid }
        currentPlanStatus: { $ref: '#/components/schemas/PlanStatus' }
        nextReviewDue: { type: [string, 'null'], format: date }
        lastTestedAt: { type: [string, 'null'], format: date-time }
        blockingIssueCount: { type: integer }
        openActionItemCount: { type: integer }
        activeRecoveryRunId: { type: [string, 'null'], format: uuid }
    ItServicePage:
      allOf:
        - $ref: '#/components/schemas/PageInfo'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/ItService' }

    # ─────────── Business Impact ───────────
...
    RecoveryObjectiveFields:
      type: object
      properties:
        scenarioId:
          type: [string, 'null']
          format: uuid
          description: null = the default objective for all scenarios
        rtoMinutes: { $ref: '#/components/schemas/Minutes' }
        rpoMinutes: { $ref: '#/components/schemas/Minutes' }
        mttrTargetMinutes: { type: [integer, 'null'], minimum: 0 }
        restorePriority: { type: integer, minimum: 1 }
        firstFunctions:
          type: array
          description: Functionality that must be available first
          items: { type: string }
    RecoveryObjectiveCreate:
      allOf:
        - $ref: '#/components/schemas/RecoveryObjectiveFields'
        - required: [rtoMinutes, rpoMinutes]
    RecoveryObjective:
      allOf:
        - $ref: '#/components/schemas/ResourceMeta'
        - $ref: '#/components/schemas/ContentMeta'
        - $ref: '#/components/schemas/RecoveryObjectiveFields'
        - type: object
          required: [rtoMinutes, rpoMinutes]
          properties:
            microserviceId: { type: string, format: uuid, readOnly: true }

    # ─────────── Recovery Strategy (Mitigation) ───────────
...